+++ processor/tests/eph.spec.ts
@@ -299,4 +299,117 @@ describe("EPH", function () {
       }
     });
   });
+
+  describe("runtimeInfo", function () {
+    it("should get hub runtime info correctly", async function () {
+      host = EventProcessorHost.createFromConnectionString(
+        EventProcessorHost.createHostName(),
+        storageConnString!,
+        EventProcessorHost.createHostName("single"),
+        ehConnString!,
+        {
+          eventHubPath: hubName!,
+          initialOffset: EventPosition.fromEnqueuedTime(Date.now())
+        }
+      );
+      const hubRuntimeInfo = await host.getHubRuntimeInformation();
+      should.equal(Array.isArray(hubRuntimeInfo.partitionIds), true);
+      should.equal(typeof hubRuntimeInfo.partitionCount, "number");
+    });
+
+    it("should get partition runtime info correctly with partitionId as string", async function () {
+      host = EventProcessorHost.createFromConnectionString(
+        EventProcessorHost.createHostName(),
+        storageConnString!,
+        EventProcessorHost.createHostName("single"),
+        ehConnString!,
+        {
+          eventHubPath: hubName!,
+          initialOffset: EventPosition.fromEnqueuedTime(Date.now())
+        }
+      );
+      const partitionInfo = await host.getPartitionInformation("0");
+      debug(">>> partitionInfo: %o", partitionInfo);
+      partitionInfo.partitionId.should.equal("0");
+      partitionInfo.type.should.equal("com.microsoft:partition");
+      partitionInfo.hubPath.should.equal(hubName);
+      partitionInfo.lastEnqueuedTimeUtc.should.be.instanceof(Date);
+      should.exist(partitionInfo.lastSequenceNumber);
+      should.exist(partitionInfo.lastEnqueuedOffset);
+    });
+
+    it("should get partition runtime info correctly with partitionId as number", async function () {
+      host = EventProcessorHost.createFromConnectionString(
+        EventProcessorHost.createHostName(),
+        storageConnString!,
+        EventProcessorHost.createHostName("single"),
+        ehConnString!,
+        {
+          eventHubPath: hubName!,
+          initialOffset: EventPosition.fromEnqueuedTime(Date.now())
+        }
+      );
+      const partitionInfo = await host.getPartitionInformation(0);
+      partitionInfo.partitionId.should.equal("0");
+      partitionInfo.type.should.equal("com.microsoft:partition");
+      partitionInfo.hubPath.should.equal(hubName);
+      partitionInfo.lastEnqueuedTimeUtc.should.be.instanceof(Date);
+      should.exist(partitionInfo.lastSequenceNumber);
+      should.exist(partitionInfo.lastEnqueuedOffset);
+    });
+
+    it("should fail getting partition information when partitionId is not a string or number", async function () {
+      host = EventProcessorHost.createFromConnectionString(
+        EventProcessorHost.createHostName(),
+        storageConnString!,
+        EventProcessorHost.createHostName("single"),
+        ehConnString!,
+        {
+          eventHubPath: hubName!,
+          initialOffset: EventPosition.fromEnqueuedTime(Date.now())
+        }
+      );
+      try {
+        await host.getPartitionInformation(false as any);
+      } catch (err) {
+        err.message.should.equal("'partitionId' is a required parameter and must be of type: 'string' | 'number'.");
+      }
+    });
+
+    it("should fail getting partition information when partitionId is empty string", async function () {
+      host = EventProcessorHost.createFromConnectionString(
+        EventProcessorHost.createHostName(),
+        storageConnString!,
+        EventProcessorHost.createHostName("single"),
+        ehConnString!,
+        {
+          eventHubPath: hubName!,
+          initialOffset: EventPosition.fromEnqueuedTime(Date.now())
+        }
+      );
+      try {
+        await host.getPartitionInformation("");
+      } catch (err) {
+        err.message.should.match(/.*The specified partition is invalid for an EventHub partition sender or receiver.*/ig);
+      }
+    });
+
+    it("should fail getting partition information when partitionId is a negative number", async function () {
+      host = EventProcessorHost.createFromConnectionString(
+        EventProcessorHost.createHostName(),
+        storageConnString!,
+        EventProcessorHost.createHostName("single"),
+        ehConnString!,
+        {
+          eventHubPath: hubName!,
+          initialOffset: EventPosition.fromEnqueuedTime(Date.now())
+        }
+      );
+      try {
+        await host.getPartitionInformation(-1);
+      } catch (err) {
+        err.message.should.match(/.*The specified partition is invalid for an EventHub partition sender or receiver.*/ig);
+      }
+    });
+  });
 });